LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/1-iframe/parent-yes-child-no-same.sub.https-expected.txt


PASS setting document.domain must give sync access
FAIL parent: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
FAIL child: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
